+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `XIT ACT`: Added "CONT Ship" and "CONT Trade" actions
 - `XIT CONTC`: Added fulfill buttons to fulfillable conditions
 - `XIT FLT`: Enhanced fleet table with detailed status, cargo, fuel, and quick actions
+- `XIT FINCH`: Added an "Equity Growth" chart
 - `XIT FLOW`: Shows daily production, consumption, and value of each material across bases
 - `XIT STO`: Per-base storage analysis
 - Added a notification count badge to the browser tab favicon (`favicon-notification-badge`)